+++ packages/eslint-plugin/src/rules/camelcase.ts
@@ -72,21 +72,26 @@ export default util.createRule<Options, MessageIds>({
      * @private
      */
     function isTSPropertyType(node: TSESTree.Node): boolean {
-      if (!node.parent) {
-        return false;
-      }
-      if (TS_PROPERTY_TYPES.includes(node.parent.type)) {
+      if (TS_PROPERTY_TYPES.includes(node.type)) {
         return true;
       }
 
-      if (node.parent.type === AST_NODE_TYPES.AssignmentPattern) {
+      if (node.type === AST_NODE_TYPES.AssignmentPattern) {
         return (
-          node.parent.parent !== undefined &&
-          TS_PROPERTY_TYPES.includes(node.parent.parent.type)
+          node.parent !== undefined &&
+          TS_PROPERTY_TYPES.includes(node.parent.type)
         );
       }
 
       return false;
+    }
+
+    function report(node: TSESTree.Identifier): void {
+      context.report({
+        node,
+        messageId: 'notCamelCase',
+        data: { name: node.name },
+      });
     }
 
     return {
@@ -103,13 +108,24 @@ export default util.createRule<Options, MessageIds>({
         }
 
         // Check TypeScript specific nodes
-        if (isTSPropertyType(node)) {
+        const parent = node.parent;
+        if (parent && isTSPropertyType(parent)) {
           if (properties === 'always' && isUnderscored(name)) {
-            context.report({
-              node,
-              messageId: 'notCamelCase',
-              data: { name: node.name },
-            });
+            report(node);
+          }
+
+          return;
+        }
+
+        if (parent && parent.type === AST_NODE_TYPES.OptionalMemberExpression) {
+          // Report underscored object names
+          if (
+            properties === 'always' &&
+            parent.object.type === AST_NODE_TYPES.Identifier &&
+            parent.object.name === node.name &&
+            isUnderscored(name)
+          ) {
+            report(node);
           }
 
           return;